requirements.txt
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
bleach==6.2.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
cyclonedx-python-lib==9.1.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.9
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.7
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
lark==1.2.2
license-expression==30.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
music21==9.7.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.2
onnxruntime==1.22.1
overrides==7.7.0
packageurl-python==0.17.5
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pillow==10.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
py-serializable==2.1.0
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
razdel==0.5.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
ruaccent==1.5.8.3
safetensors==0.6.2
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
tornado==6.5
tqdm==4.66.3
traitlets==5.14.3
transformers==4.55.4
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
uri-template==1.3.0
urllib3==2.5.0
virtualenv==20.26.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
